+++ bin/extract-ORFs.py
@@ -68,11 +68,16 @@ if __name__ == '__main__':
         try:
             for translation in translations(read):
                 for orf in translation.ORFs():
+                    # Omit open ORFs if they're unwanted.
+                    if ((orf.openLeft or orf.openRight) and not allowOpenORFs):
+                        continue
+
+                    # Check the length requirements, if any.
                     if ((minORFLength is None or len(orf) >= minORFLength) and
-                        (maxORFLength is None or len(orf) <= maxORFLength)
-                        ) or (allowOpenORFs and (
-                            orf.openLeft or orf.openRight)):
+                            (maxORFLength is None or
+                                len(orf) <= maxORFLength)):
                         write(orf.toString('fasta'))
+
         except TranslationError as error:
             print('Could not translate read %r sequence %r (%s).' %
                   (read.id, read.sequence, error), file=sys.stderr)